Moved RSSPostManager to helpers
