+++ src/helpers/htmlMetaData.ts
@@ -1,5 +1,5 @@
 import { OpenGraphData, getHtmlOpenGraphData } from './openGraph';
-import { RSSFeedManager } from '../RSSPostManager';
+import { RSSFeedManager } from './RSSPostManager';
 import { fetchSiteFaviconUrl } from './favicon';
 import { HtmlUtils } from './HtmlUtils';
 import { Feed } from '../models/Feed';
